Fixed variable type in npc.cpp (#3996)

+++ src/enums.h
@@ -524,12 +524,12 @@ struct LightInfo {
 struct ShopInfo {
 	uint16_t itemId = 0;
 	int32_t subType = 1;
-	uint32_t buyPrice = 0;
-	uint32_t sellPrice = 0;
+	int64_t buyPrice = 0;
+	int64_t sellPrice = 0;
 	std::string realName = "";
 
 	ShopInfo() = default;
-	ShopInfo(uint16_t itemId, int32_t subType = 0, uint32_t buyPrice = 0, uint32_t sellPrice = 0, std::string realName = "")
+	ShopInfo(uint16_t itemId, int32_t subType = 0, int64_t buyPrice = 0, int64_t sellPrice = 0, std::string realName = "")
 		: itemId(itemId), subType(subType), buyPrice(buyPrice), sellPrice(sellPrice), realName(std::move(realName)) {}
 };
